Only read melange configs in the directory the command is being executed from
This was the original behavoiur but changed in https://github.com/wolfi-dev/wolfictl/pull/31/files to handle sub folders containing melange configs.

We decided not to take this approach and moved the configs out into the root.

Instead if a melange config exists in a sub folder then lint and update commands should run in these specific folders.

Signed-off-by: James Rawlings <jrawlings@chainguard.dev>