Don't explicitly check CreatesCycle
This cuts NewGraph in half for me (12s -> 6s).

Before we call AddEdge, we always call CreatesCycle.

The first thing that AddEdge does is call CreatesCycle, which is
currently 99% of the execution time of AddEdge.

Instead of calling that twice for every edge, just call AddEdge
and check to see if it would have created a cycle.

Because we very rarely hit cycles, 99% of the time, this is just
duplicated work.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>