Consider subpackage dependencies in wolfictl text
We previously ignored subpackages because we can't build them separately
from their containing package, but filtering out subpackages means that
the dependency graph is incomplete.

Instead of filtering them out before calling the toposort, we just
omit subpackages when we emit the build order. Given that subpackages
always depend on their parents, the toposort will still be valid.

Before this change, the first package was "wget", which has a dependency
on "openssl-dev", which doesn't exist in the stage3 bootstrap repo.

After this change, the first package is "openssl", which seems correct.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>