Use sync.Cond over context cancellation
sync.Cond was basically made for this situation where we have a bunch of
goroutines that we want to park until something else finishes.

I've also cleaned up a little bit of how we propagate errors and logs so
that cancelling the program results in useful output and not logspam.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>