build: Support passing packages as argv
This makes wolfictl build work as a replacement for "make package/foo".

I've also dropped an errgroup we use for building every arch
simultaneously. I think it kind of betrays the -j flag a bit, and doing
that made it easier to add the per-arch logfile to the error message
that gets returned.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>