build: Group by arch
We used to do something like this:

```
for pkg in packages():
  for arch in archs:
    build(pkg, arch)
```

This change looks more like this:

```
for arch in archs:
  for pkg in packages(arch):
    build(pkg, arch)
```

This works a little bit better because the dependency graph only makes
sense within a given arch; e.g. we shouldn't wait for aarch64/go to
finish before building x86_64/crane.

This also tees us up to prune the set of packages we need to build based
on an existing APKINDEX.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>