Index APKs if they are already built
Instead of skipping the whole thing and only indexing/updating orphaned
APKs at the end, we will adopt any built-but-not-indexed APKs into our
local APKINDEX so that we can use them in the current build.

This will allow us to recover in the case where only certain packages
succeed to build (due to cyclical dependencies or otherwise) for new
packages where we can't fall back to a previous version.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>