fix(scan): enable pseudoversion comparison
After my Grype PR (https://github.com/anchore/grype/pull/1797), Anchore added a follow-up PR that makes this behavior cnofigurable. It is on by default for the Grype CLI but not for the Grype library. This commit enables it for the Grype library so we can see this kind of match from wolfictl scans.

Signed-off-by: Dan Luhring <dluhring@chainguard.dev>