bundles: Make traces slightly more organized
This disables the otelhttp traces that the GCS client was emitting,
which are probably useful but don't seem to respect the parent trace
context, so they are just top level noise.

This also adds a "start <pkg>" trace to organize the "waiting for <dep>"
traces under a reasonable parent instead of the root.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>