bundles: Embed build date epoch in bundles
Originally attempted to do this by just shelling out to git for each
file individually. For a small number of recently modified files, this
isn't too bad, but for build-the-world builds, this ends up taking a
very long time because we have to traverse the commit graph for each
melange config, which is 2400 times, and that can take ~1s per file.

Instead, we traverse the commit graph once by shelling out to git and
just parsing the output for every changed file per non-merge commit.

This takes about 2s on my machine, whereas doing it once per file took
about 20 minutes.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>