bundles: Handle multiple builds (best-effort)
If a pod has already been scheduled for a given APK, we'll just watch it
instead of creating a duplicate pod. This avoids duplicating ongoing
builds. Every controller that is watching a pod will still attempt to
upload it, which gives us some resilience re: failed controllers, but
requires the APK repo to handle concurrent uploads gracefully.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>