feat(updatebot)!: remove update bot as not maintained anymore
BREAKING CHANGE: The updatebot code is old and has since been improved to run well on GCP rather than GitHub Actions.  This is to help with stability, monitoring, alerting and speed ad which new features can be developed to support the growing needs and scale of Wolfi.

This code is not being maintained anymore and thus will become a burden on the Wolfi maintainers which Wolfictl is primarily developed and supported.

The update check remains for now which is why some code still exists relating to updates.  This will likely be removed in the future too.

Signed-off-by: James Rawlings <jrawlings@chainguard.dev>