Consider subpackage runtime deps for graph
We were only looking at main package runtime deps when constructing the
dependency graph, which meant we missed a lot of edges that were very
important.

In order for this to work, I had to re-work some of the logic in
Targets(), which flattens subpackages into their parents. Doing that
re-introduced cycles because it previously wasn't ignoring non-local
dependencies (oops).

Once we eliminate the actual cycles in wolfi, we should be in a better
place.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>