text: Only consider local repo for self-providers
This was seemingly added to prevent accidentally having cycles in the
dependency graph, but this also prevents us from retiring older versions
of packages such that we only have the most recent version. This matters
if, for example, we wanted to migrate to a new repository location
without backfilling every package ever.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>