lint/rulest_test: remove check for error and want error
This check prevents failuers from being printed by the assertion which
is useful when developing tests.

```
go test ./pkg/lint/
--- FAIL: TestLinter_Rules (0.01s)
    --- FAIL: TestLinter_Rules/invalid-spdx-license.yaml (0.00s)
        rules_test.go:480: Lint() error = <nil>, wantErr true
FAIL
FAIL	github.com/wolfi-dev/wolfictl/pkg/lint	0.411s
FAIL
```

vs

```
go test ./pkg/lint/
--- FAIL: TestLinter_Rules (0.01s)
    --- FAIL: TestLinter_Rules/invalid-spdx-license.yaml (0.00s)
        rules_test.go:495:
            	Error Trace:	/Users/justin/work/wolfictl/pkg/lint/rules_test.go:495
            	Error:      	Not equal:
            	           	expected: &errors.errorString{s:"[valid-spdx-license]: license \"Apache License 2.0\" is not valid SPDX license (ERROR)"}
            	           	actual  : &errors.errorString{s:"[invalid-spdx-license]: valid-spdx-license \"Apache License 2.0\" not valid SPDX license (ERROR)"}

            	           	Diff:
            	           	--- Expected
            	           	+++ Actual
```

The condition it's checking for is a bit strange. I read it as "if
there is an error in the linting process" AND "the state of there being an error
is the same as whether or not we want an error."  This isn't super
usefull to us.  We don't really care if the linter itself failed and
we're not counting the failures in the linting processes we're counting the
failures from a specific linter we ran. If the linting process itself
failed. We won't get the correct result object back to match in the
lower comparisons where we'll get details about expected vs actual
errors.  Instead we'll just assert that the linting process produced no
failures.