adv: introduce the new advisory-schema module
This commit introduces the new `advisory-schema` module which contains the
advisory schema and validation logic extracted from wolfictl.

It ensures a clear separation between the codebase that defines the schema
and the codebases that consumes it, such as this `wolfictl`
repository and others.

Note - the initial release version of `advisory-schema` (v0.37.0) is aligned with the
upcoming `wolfictl` release that first depends on this module, making the refactoring
history easier to trace. Each project will  follow its own release
schedule moving forward.

Signed-off-by: Maxime Gréau <maxime.greau@chainguard.dev>