lint: add rule to make sure update identifier matches git repo
The update identifier must match the github repository because, otherwise,
updating the package fails. Actualy, we have had a case already where a package
definition was copied & pasted from another and these values were not updated
to match.

To catch any further mismatches like that this commit adds a new linter rule:
update-identifier-must-match-git-repository.

It checks, in a case insensitive manner, whether the GitHub repository URL
contains a GitHub identifier that matches the one from the update section of
the package yaml.

Furthermore, because a single package can have multiple pipelines, this linter
looks for a single match, i.e. a single repository URL from all of the
pipelines that are using git-checkout must match the update identifier in order
to pass.

Note that is possible to opt-out from it by using the #nolint feature with this
linter name. See:
update-identifier-not-matching-git-checkout-repository-nolint.yaml to see how
that is done.

Signed-off-by: Konrad Zapałowicz <konrad.zapalowicz@chainguard.dev>