chore(sbom): resolve warning for directory source
Syft logs a warning if you're using a directory source but don't set a 'source.Alias' value for the source. This log message isn't warning about anything that impacts the accuracy of the analysis, but it's nonetheless noisy and annoying. This PR plugs in an alias value to avoid the warning. It also moves the Syft-to-clog wiring slightly earlier to make the effect of this change reproducible with the local CLI.

Signed-off-by: Dan Luhring <dluhring@chainguard.dev>