bump.go: Ignore bot-generated CVE/GHSA comments
octo-sts will [add CVE/GHSA
comments](https://github.com/wolfi-dev/os/commit/352a0137b9f0f66184807c9a10cd451eea094961) whenever it
bumps a package's epoch to fix a security issue.  This is good, but we
should remove those comments when we invoke `wolfictl bump`, otherwise
they will wrongly imply that the manual bump was related to the
security fix as well.

I tried to be conservative here and only remove the comments that
start with "(CVE|GHSA)-", because there is a possibility to find an
actual human-made comment and we don't want to touch those (although
IMO it should be considered bad practice to leave manual comments
right after the `epoch` definition).

Signed-off-by: Sergio Durigan Junior <sergiodj@chainguard.dev>